made WeatherDateTemperature.swift and moved it here
